fix: automatically migrate old configuration file
Release 0.13.0 added a new configuration field, which prevented the
integration to load old configuration files.
If the new `name` field is missing, automatically scan the device and
update the configuration. This is done during startup and also at the
start of a reconfiguration, since the network might not yet be available
when the integration starts.