feat: New global volume management  (#37)
Add support to set volume with level.
Also the volume is applied to all connected devices unless the user unchecks it in the setup flow/reconfigure flow.
By default this new setting is enabled. Also the reported volume level is now updated with an average value of all volume levels.